sla_fusion_backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
pandas==2.0.1
openpyxl==3.1.2
xlsxwriter==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
python-magic==0.4.27
aiofiles==23.1.0
pydantic==1.10.7
python-multipart==0.0.6
boto3==1.26.158  # For AWS S3 integration (optional)
google-cloud-storage==2.9.0  # For GCP Storage (optional)
python-magic-bin==0.4.14; sys_platform == 'win32'  # Windows specific dependency
